requirements.txt
# === Core: Google ADK ===
google-adk[extensions]>=1.2.0

# === Database ===
asyncpg>=0.29.0

# === API Server ===
fastapi>=0.115.0
uvicorn>=0.32.0

# === Environment ===
python-dotenv>=1.0.0

# === Observability ===
arize-phoenix>=8.0.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0
openinference-instrumentation-google-adk>=0.1.0

# === Evaluation (LLM Judge via Ollama) ===
requests>=2.31.0

# === Testing ===
pytest>=8.0.0
pytest-asyncio>=0.24.0
